fix(aionrs): contain unix exec timeouts
Move the process runner implementation out of lib.rs and keep the crate root as a thin module entrypoint.

On Unix, spawn commands into their own process group and kill the group on timeout so shell descendants are not left running.

Cover the process-group timeout behavior with a regression test.
